requirements.txt
azure-storage-queue 
azure-storage-blob 
azure-identity
requests